# Conflicts:
#	cmd/node/factory/structs.go
#	cmd/node/factory/structsStatusHandlers.go
#	cmd/node/main.go
#	data/trie/trieStorageManager.go
#	integrationTests/consensus/testInitializer.go
#	integrationTests/mock/blockProcessorMock.go
#	integrationTests/testProcessorNode.go
#	integrationTests/testSyncNode.go
#	node/node.go
#	node/node_test.go
#	process/sync/argBootstrapper.go
#	process/sync/baseSync.go
#	process/sync/metablock.go
#	process/sync/metablock_test.go
#	process/sync/shardblock.go
#	process/sync/shardblock_test.go

+++ data/trie/trieStorageManager.go
@@ -38,7 +38,7 @@ type trieStorageManager struct {
 	snapshotReq        chan *snapshotsQueueEntry
 	pruningBuffer      atomicBuffer
 	snapshotInProgress uint32
-	maxSnapshots       uint8
+	maxSnapshots       uint32
 	cancelFunc         context.CancelFunc
 
 	dbEvictionWaitingList data.DBRemoveCacher
@@ -448,7 +448,7 @@ func (tsm *trieStorageManager) getSnapshotDb(newDb bool) data.DBWriteCacher {
 		return nil
 	}
 
-	if uint8(len(tsm.snapshots)) > tsm.maxSnapshots {
+	if uint32(len(tsm.snapshots)) > tsm.maxSnapshots {
 		tsm.removeSnapshot()
 	}
